implement `sanitize_selection` for `Policy` and `PluginPolicy`
