+++ src/lib.rs
@@ -8,6 +8,8 @@ pub mod plugin_policy;
 pub mod policy;
 pub mod traits;
 
+pub(crate) mod macros;
+
 #[doc(inline)]
 pub use directives::{Permissive, Restrictive};
 pub use policy::*;
